tabel ok

+++ database/migrations/2020_08_21_140815_create_parameters_table.php
@@ -15,7 +15,7 @@ class CreateParametersTable extends Migration
     {
         Schema::create('parameters', function (Blueprint $table) {
             $table->uuid('id')->primary();
-            $table->integer('min_support');
+            $table->float('min_support');
             $table->integer('min_confidence');
         });
     }
